\name{RSquamataBase-package}
\alias{RSquamataBase-package}
\alias{RSquamataBase}
\docType{package}
\title{
\packageTitle{RSquamataBase}
}
\description{
\packageDescription{RSquamataBase}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RSquamataBase}
\packageIndices{RSquamataBase}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RSquamataBase}

Maintainer: \packageMaintainer{RSquamataBase}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
